make version number optional on archive command
Also includes tests for archive command with one that's skipped because of some weird behavior. It's possible the feature is broken but I have a feeling it's related to the stubbing. Will need to investigate further as to what the cause of the failure is. Mainly this means testing on a real project to verify the behavior works as expected and then take the time to fix the test.