check for locationType when server is being started up rather than any cli command being executed
This fixes the recent issues where it was forcing you to set it before anything
would work. Even the generator that sets it for you. This way it only does it
when the server is actually running and assets are being served.

This also makes it so that the new `ember install:addon` command will work with
this because it was preventing that from working